refactor volumetrics folder
