--- nanome/_internal/volumetric/serialization/volume_properties_serializer.py
@@ -1,25 +1,0 @@
-from .. import _VolumeData
-from nanome._internal.util.type_serializers import TypeSerializer, ArraySerializer
-from . import _VolumeLayerSerializer
-
-
-class _VolumePropertiesSerializer(TypeSerializer):
-    def __init__(self):
-        self.__array = ArraySerializer()
-        self.__array.set_type(_VolumeLayerSerializer())
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "VolumeProperties"
-
-    def serialize(self, version, value, context):
-        context.write_bool(value._visible)
-        context.write_bool(value._boxed)
-        context.write_bool(value._use_map_mover)
-        context.write_int(int(value._style))
-        context.write_using_serializer(self.__array, value._layers)
-
-    def deserialize(self, version, context):
-        raise NotImplementedError
